refactor: reorganize service and component imports, remove unused files
- Deleted index files for components and services to streamline imports.
- Updated imports in content, options, and popup entry points to directly reference specific service files.
- Removed unused utility files related to season configuration and storage management.
- Updated package version to 1.2.15 and adjusted dependencies in package.json and yarn.lock.
