--- components/index.ts
@@ -1,3 +1,0 @@
-// Export all components for easy importing
-export { default as UIComponents } from "./uiComponents";
-export { default as AccountUIService } from "./accountUIComponents";
