+++ package.json
@@ -2,7 +2,7 @@
   "name": "google-cloud-skills-boost-helper",
   "description": "__MSG_extDescription__",
   "private": true,
-  "version": "1.2.14",
+  "version": "1.2.15",
   "author": "dev@eplus.dev",
   "license": "MIT",
   "homepage_url": "https://eplus.dev",
@@ -22,15 +22,12 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.3.1",
-    "@wxt-dev/storage": "^1.1.1",
     "axios": "^1.18.0",
     "firebase": "^12.15.0",
     "fuse.js": "^7.4.2",
-    "intro.js": "^8.3.2",
     "marked": "^18.0.5",
     "react": "^19.2.7",
     "react-dom": "^19.2.7",
-    "rxjs": "^7.8.2",
     "tailwindcss": "^4.0.16"
   },
   "devDependencies": {
@@ -41,13 +38,12 @@
     "@vitejs/plugin-react": "^6.0.1",
     "@vitejs/plugin-react-swc": "^4.3.0",
     "@vitest/coverage-v8": "^4.1.0",
-    "@wxt-dev/module-react": "1.2.2",
     "jsdom": "^29.0.1",
-    "typescript": "^5.6.3",
     "tmp": "^0.2.6",
+    "typescript": "^5.6.3",
     "vite": "^8.0.16",
     "vitest": "^4.1.0",
-    "wxt": "^0.20.26"
+    "wxt": "^0.20.27"
   },
   "resolutions": {
     "uuid": "file:vendor/uuid",
