--- services/index.ts
@@ -1,16 +1,0 @@
-// Export all services for easy importing
-export { default as ApiClient } from "./apiClient";
-export { default as SearchService } from "./searchService";
-export { default as LabService } from "./labService";
-export { default as ProfileService } from "./profileService";
-export { default as ArcadeApiService } from "./arcadeApiService";
-export { default as StorageService } from "./storageService";
-export { default as PopupUIService } from "./popupUIService";
-export { default as BadgeService } from "./badgeService";
-export { default as PopupService } from "./popupService";
-export { default as OptionsService } from "./optionsService";
-export { default as MarkdownService } from "./markdownService";
-export { default as AccountService } from "./accountService";
-export { default as TourService } from "./tourService";
-export * as BrowserService from "./browserService";
-export { default as ExportService } from "./exportService";
